+++ client/cody-shared/src/chat/markdown.ts
@@ -35,6 +35,7 @@ const DOMPURIFY_CONFIG = {
         's',
         'u',
     ],
+    ALLOWED_URI_REGEXP: /^vscode:?\/\/[^\s#$./?].\S*$/i,
 }
 
 /**